Move trace_io contents to trace_common
